% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RappPKPLT-package.R
\docType{package}
\name{RappPKPLT-package}
\alias{RappPKPLT}
\alias{RappPKPLT-package}
\title{RappPKPLT: R Shiny Application for a PK-platelet model}
\description{
An R Shiny application for interactive simulation using a PK-platelet model.
}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com}

}
\keyword{internal}
