% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run Shiny application for PK-platelet-GDF15 simulation}
\usage{
app()
}
\value{
Nothing
}
\description{
Run Shiny application for interactive PKPD simulation using a PK-platelet-GDF15 model.
The simulation is powered by the \link[mrgsolve:mrgsolve_package]{mrgsolve::mrgsolve} package.
}
\examples{
\dontrun{
app()
}
}
