% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RappPKPLT-package.R
\docType{package}
\name{RappPKPLT-package}
\alias{RappPKPLT}
\alias{RappPKPLT-package}
\title{RappPKPLT: Shiny Application for a PK-Platelet Model}
\description{
A Shiny application for interactive PKPD simulation using a PK-platelet model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soutomas/RappPKPLT}
  \item Report bugs at \url{https://github.com/soutomas/RappPKPLT/issues}
}

}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com}

}
\keyword{internal}
